==> 01_overall pattern.Rmd <==
```{r echo=FALSE, warning=FALSE, message=FALSE}
# Library
library(ggplot2)
library(tidyverse)
library(janitor)
```


```{r echo=FALSE, warning=FALSE, message=FALSE}
datasepa_data <- read_csv("data_preparation/sepdata_departure.csv") 
```

```{r}
countdata1 <- datasepa_data%>%
  group_by(date) %>%
  summarise(trip_per_day=sum(count),
            average_duration=(mean(duration)))
```

```{r}
datasepa_month <- datasepa_data %>% 
  mutate(midday = count*15)
datasepa_month <- unite (datasepa_month, "monthdata", year, month, midday, sep = "-", remove = FALSE)
datasepa_month$monthdata <- as.Date(datasepa_month$monthdata)

countdata2 <- datasepa_month%>%
  group_by(monthdata) %>%
  summarise(trip_monthly=sum(count)/30,
            average_duration=(mean(duration)))
```


```{r}
datebreaks <- seq(as.Date("2019-01-01"), as.Date("2021-03-31"),
                  by = "3 month")
ggplot(data = countdata1, mapping = aes(x = date, y = trip_per_day)) +
  geom_point(aes(size = average_duration, alpha = 0.002))+
  geom_line(data = countdata2, 
            mapping = (aes(x = monthdata, 
                           y = trip_monthly,
                           color = 'monthly average')),
            size = 1.3)+
  xlab("Date")+
  ylab("Number of bikeshare ridership (*1000)")+
  scale_x_date(breaks = datebreaks)+
  scale_y_continuous(labels = function(y) y / 1000)+
  ggtitle("Trend of bicycle hire uptake in London")+
  theme(axis.title = element_text(size=18),axis.text = element_text(size=16),
        strip.text = element_text(size=18))+
  theme_classic() 
```
```{r}
ggsave("pic/overall uptake.jpg", width = 15, height = 6)
```
