==> 08_community detection.Rmd <==
## Community detection

```{r echo=FALSE, warning=FALSE, message=FALSE}
# Library
library(tidyverse)
library(janitor)
library (readr)
library(igraph)
```
```{r echo=FALSE, warning=FALSE, message=FALSE}
cycle <- read_csv("data_preparation/stations.csv") %>% 
  select(., c(1, 3, 4))
```

```{r echo=FALSE, warning=FALSE, message=FALSE}
datasepa_data <- read_csv('data_preparation/sepdata_departure.csv')
```

## Function for extracting and calculating data for month 

```{r}
countdatafun <- function(data, m, y, n1, n2){
  alldatamonth <- data %>%
    filter(month == m) #'03'
  
  alldatamonth$day <- as.numeric(alldatamonth$day)
  monthlast <- alldatamonth %>% 
    filter(day >= n1) %>% 
    filter(day <= n2)
  

  monthlastyear <- monthlast %>%
    filter(year == y)
  
  countmonyear <- monthlastyear %>%
    group_by(end_station_id, start_station_id) %>%
    summarise(route=sum(count))
  
  return(countmonyear)
 
}
```

## calculation for weekdays

```{r echo=FALSE, warning=FALSE, message=FALSE}

countmar2019 <- countdatafun(datasepa_data, "03", "2019", 25, 29)
countmay2019 <- countdatafun(datasepa_data, "05", "2019", 20, 24)
countaug2019 <- countdatafun(datasepa_data, "08", "2019", 19, 23)
countmar2020 <- countdatafun(datasepa_data, "03", "2020", 23, 27)
countmay2020 <- countdatafun(datasepa_data, "05", "2020", 25, 29)
countaug2020 <- countdatafun(datasepa_data, "08", "2020", 24, 28)
countmar2021 <- countdatafun(datasepa_data, "03", "2021", 22, 26)
```



```{r echo=FALSE, warning=FALSE, message=FALSE}
countmay2019 <- countmay2019 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
countmay2020 <- countmay2020 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
countaug2019 <- countaug2019 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
countaug2020 <- countaug2020 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")

countmar2019 <- countmar2019 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
countmar2020 <- countmar2020 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
countmar2021 <- countmar2021 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
```


### Function for community detection

```{r}
connectfun <- function(countmonth){

  listid <- cycle$id
  countmonth <- countmonth[countmonth$source %in% listid,]
  countmonth <- countmonth[countmonth$target %in% listid,]
  
  comdetect<-graph_from_data_frame(countmonth, directed = FALSE, vertices = cycle)
  # Assigning edge attribute to each edge
  E(comdetect)$weight<-E(comdetect)$freq 
  
  # Louvain clustering
  clusters <- cluster_louvain(comdetect)
  # communities(clusters) # check vertices belonging to different clusters
  communitylist <- communities(clusters)
  
  return(communitylist)
}
```

### Function for QGIS visualization

```{r}
gisfun <- function(countmonth, csvname){
  communitylist <- connectfun(countmonth)
  df <- data.frame(communitylist)
  colnames(df) <- c("map")
  listlen <- list()
  for (i in df$map){
    list1 <- list(r = i)
    listlen <- c(listlen, list(length = length(list1$r)))
  }
  
  dfcom <- t(data.frame(listlen)) %>% 
    cbind(df, .) %>% 
    rename("length" = ".") %>% 
    filter(length > 1)

  count = 0
  listr1 <- list()
  listr2 <- list()
  for (i in dfcom$map){
    listr1 <- append(listr1, i)
  }
  for (i in dfcom$length){
    count = count + 1
    sublist <- rep(paste("cluster", count), times = i)
    listr2 <- append(listr2, sublist)
  }  
  connection1 <- t(data.frame(listr1)) 
  connection2 <- t(data.frame(listr2))
  connection <- cbind(connection1, connection2)
  row.names(connection) <- NULL
  colnames(connection) <- c("id","cluster") 
  
  connection <- data.frame(connection)
  connection$id <- as.numeric(connection$id)
  gisconnect <- connection %>% 
    left_join(., cycle,
              by = c("id" = "id"))
  
  write.table(gisconnect,csvname,row.names=FALSE,col.names=TRUE,sep=",")
  return(gisconnect)
}
```

```{r echo=FALSE, warning=FALSE, message=FALSE}
gismay2019 <- gisfun(countmay2019, "table/qgis_community_detection/weekdays/gismay2019.csv")
gismay2020 <- gisfun(countmay2020, "table/qgis_community_detection/weekdays/gismay2020.csv")
gisaug2019 <- gisfun(countaug2019, "table/qgis_community_detection/weekdays/gisaug2019.csv")
gisaug2020 <- gisfun(countaug2020, "table/qgis_community_detection/weekdays/gisaug2020.csv")
gismar2019 <- gisfun(countmar2019, "table/qgis_community_detection/weekdays/gismar2019.csv")
gismar2020 <- gisfun(countmar2020, "table/qgis_community_detection/weekdays/gismar2020.csv")
gismar2021 <- gisfun(countmar2021, "table/qgis_community_detection/weekdays/gismar2021.csv")
```

## calculation for weekends

```{r echo=FALSE, warning=FALSE, message=FALSE}

countmar2019 <- countdatafun(datasepa_data, "03", "2019", 30, 31)
countmay2019 <- countdatafun(datasepa_data, "05", "2019", 25, 26)
countaug2019 <- countdatafun(datasepa_data, "08", "2019", 24, 25)
countmar2020 <- countdatafun(datasepa_data, "03", "2020", 28, 29)
countmay2020 <- countdatafun(datasepa_data, "05", "2020", 30, 31)
countaug2020 <- countdatafun(datasepa_data, "08", "2020", 29, 30)
countmar2021 <- countdatafun(datasepa_data, "03", "2021", 27, 28)
```

```{r echo=FALSE, warning=FALSE, message=FALSE}
countmay2019 <- countmay2019 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
countmay2020 <- countmay2020 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
countaug2019 <- countaug2019 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
countaug2020 <- countaug2020 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")

countmar2019 <- countmar2019 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
countmar2020 <- countmar2020 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
countmar2021 <- countmar2021 %>% 
  rename("source"="start_station_id",
         "target"="end_station_id",
         "freq" = "route")
```

### Function for community detection

```{r}
connectfun <- function(countmonth){

  listid <- cycle$id
  countmonth <- countmonth[countmonth$source %in% listid,]
  countmonth <- countmonth[countmonth$target %in% listid,]
  comdetect<-graph_from_data_frame(countmonth, directed = FALSE, vertices = cycle)
  # Assigning edge attribute to each edge
  E(comdetect)$weight<-E(comdetect)$freq 
  
  # Louvain clustering
  clusters <- cluster_louvain(comdetect)
  # communities(clusters) # check vertices belonging to different clusters
  communitylist <- communities(clusters)
  return(communitylist)
}
```

### Function for QGIS visualization

```{r}
gisfun <- function(countmonth, csvname){
  communitylist <- connectfun(countmonth)
  df <- data.frame(communitylist)
  colnames(df) <- c("map")
  listlen <- list()
  for (i in df$map){
    list1 <- list(r = i)
    listlen <- c(listlen, list(length = length(list1$r)))
  }
  
  dfcom <- t(data.frame(listlen)) %>% 
    cbind(df, .) %>% 
    rename("length" = ".") %>% 
    filter(length > 1)

  count = 0
  listr1 <- list()
  listr2 <- list()
  for (i in dfcom$map){
    listr1 <- append(listr1, i)
  }
  for (i in dfcom$length){
    count = count + 1
    sublist <- rep(paste("cluster", count), times = i)
    listr2 <- append(listr2, sublist)
  }  
  connection1 <- t(data.frame(listr1)) 
  connection2 <- t(data.frame(listr2))
  connection <- cbind(connection1, connection2)
  row.names(connection) <- NULL
  colnames(connection) <- c("id","cluster") 
  
  connection <- data.frame(connection)
  connection$id <- as.numeric(connection$id)
  gisconnect <- connection %>% 
    left_join(., cycle,
              by = c("id" = "id"))
  
  write.table(gisconnect,csvname,row.names=FALSE,col.names=TRUE,sep=",")
  return(gisconnect)
}
```

```{r echo=FALSE, warning=FALSE, message=FALSE}
gismay2019 <- gisfun(countmay2019, "table/qgis_community_detection/weekends/gismay2019.csv")
gismay2020 <- gisfun(countmay2020, "table/qgis_community_detection/weekends/gismay2020.csv")
gisaug2019 <- gisfun(countaug2019, "table/qgis_community_detection/weekends/gisaug2019.csv")
gisaug2020 <- gisfun(countaug2020, "table/qgis_community_detection/weekends/gisaug2020.csv")
gismar2019 <- gisfun(countmar2019, "table/qgis_community_detection/weekends/gismar2019.csv")
gismar2020 <- gisfun(countmar2020, "table/qgis_community_detection/weekends/gismar2020.csv")
gismar2021 <- gisfun(countmar2021, "table/qgis_community_detection/weekends/gismar2021.csv")
```
